+++ src/test/slow/entry-point.test.ts
@@ -1,7 +1,6 @@
 import { tempdirProject } from "@typestrong/fs-fixture-builder";
 import { deepStrictEqual as equal, ok } from "assert";
 import { join } from "path";
-import { tmpdir } from "os";
 import { Application, EntryPointStrategy, TSConfigReader } from "../..";
 
 const fixture = tempdirProject();
@@ -73,94 +72,5 @@ describe("Entry Points", () => {
             1,
             "entry-points/index.ts should have been the sole entry point",
         );
-    });
-
-    it("Supports resolving packages", () => {
-        app.bootstrap({
-            tsconfig: tsconfig,
-            entryPoints: [fixture.cwd],
-            entryPointStrategy: EntryPointStrategy.LegacyPackages,
-        });
-
-        const entryPoints = app.getEntryPoints();
-        ok(entryPoints);
-        equal(entryPoints.length, 1);
-        equal(entryPoints[0].version, void 0);
-        equal(entryPoints[0].readmeFile, void 0);
-    });
-
-    it("Supports resolving packages outside of cwd", () => {
-        const fixture = tempdirProject({ rootDir: tmpdir() });
-        fixture.addJsonFile("tsconfig.json", {
-            include: ["."],
-        });
-        fixture.addJsonFile("package.json", {
-            main: "index.ts",
-        });
-        fixture.addFile("index.ts", "export function fromIndex() {}");
-        fixture.write();
-
-        app.bootstrap({
-            tsconfig: tsconfig,
-            entryPoints: [fixture.cwd],
-            entryPointStrategy: EntryPointStrategy.LegacyPackages,
-        });
-
-        const entryPoints = app.getEntryPoints();
-        fixture.rm();
-        ok(entryPoints);
-        equal(entryPoints.length, 1);
-        equal(entryPoints[0].version, void 0);
-    });
-
-    it("Supports custom tsconfig files #2061", () => {
-        const fixture = tempdirProject({ rootDir: tmpdir() });
-        fixture.addJsonFile("tsconfig.lib.json", {
-            include: ["."],
-        });
-        fixture.addJsonFile("package.json", {
-            main: "index.ts",
-            typedoc: {
-                tsconfig: "tsconfig.lib.json",
-            },
-        });
-        fixture.addFile("index.ts", "export function fromIndex() {}");
-        fixture.write();
-
-        app.bootstrap({
-            tsconfig: tsconfig,
-            entryPoints: [fixture.cwd],
-            entryPointStrategy: EntryPointStrategy.LegacyPackages,
-        });
-
-        const entryPoints = app.getEntryPoints();
-        fixture.rm();
-        ok(entryPoints);
-        equal(entryPoints.length, 1);
-    });
-
-    it("Supports automatically discovering the readme files", () => {
-        const fixture = tempdirProject();
-        fixture.addJsonFile("tsconfig.json", {
-            include: ["."],
-        });
-        fixture.addJsonFile("package.json", {
-            main: "index.ts",
-        });
-        fixture.addFile("reaDME.md", "Text");
-        fixture.addFile("index.ts", "export function fromIndex() {}");
-        fixture.write();
-
-        app.bootstrap({
-            tsconfig: tsconfig,
-            entryPoints: [fixture.cwd],
-            entryPointStrategy: EntryPointStrategy.LegacyPackages,
-        });
-
-        const entryPoints = app.getEntryPoints();
-        fixture.rm();
-        ok(entryPoints);
-        equal(entryPoints.length, 1);
-        equal(entryPoints[0].readmeFile, join(fixture.cwd, "reaDME.md"));
     });
 });